Changes to Annotations to support Poly Qual API
Added param() to all units annotations to make it work with the poly version of Qual API.
Copied and moved all of these annotations to a new qual directory to keep it separate and distinct from the classic units checker.
Designed a UnitsQualPool class which loads the qualifiers via reflection and maintains a singleton list of unique qualifiers. Supports dynamic creation of any qualifier that doesn't have a pre-existing annotation (eg gigga-meters).
Additional bug fixes.