Scalar number literals change
Changed rule to allow scalar number literals to be passed as an argument to any method parameter that has a unit other than unitsbottom without error, as those literals are tied to those method calls and can be safely assumed to take on the unit of the parameter. Scalar variables passed to such methods still result in errors.